Restore example and add one for the visitor
# Conflicts:
#	Polygon_mesh_processing/examples/Polygon_mesh_processing/triangulate_faces_example.cpp
